refactor(toggle): rename component for docs sake
