+++ src/components/Toggle/index.js
@@ -1,2 +1,2 @@
-export { default as MToggle } from './src/ToggleFormItem.vue';
+export { default as MToggle } from './src/Toggle.vue';
 export { default as MToggleControl } from './src/ToggleControl.vue';
